requirements.txt
boto3==1.24.14
botocore==1.27.14
certifi==2022.6.15
charset-normalizer==2.0.12
Faker==13.13.0
idna==3.3
jmespath==1.0.1
kafka-python==2.0.2
python-dateutil==2.8.2
requests==2.28.0
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.9
